+++ modules/calib3d/test/test_undistort_badarg.cpp
@@ -40,7 +40,7 @@
 //M*/
 
 #include "test_precomp.hpp"
-#include "opencv2/imgproc/imgproc_c.h"
+#include "opencv2/calib3d/calib3d_c.h"
 
 namespace opencv_test { namespace {
