+++ samples/android/camera-calibration/src/org/opencv/samples/cameracalibration/OnCameraFrameRender.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.opencv.android.CameraBridgeViewBase.CvCameraViewFrame;
+import org.opencv.calib3d.Calib3d;
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
 import org.opencv.core.MatOfPoint;
@@ -50,7 +51,7 @@ class UndistortionFrameRender extends FrameRender {
     @Override
     public Mat render(CvCameraViewFrame inputFrame) {
         Mat renderedFrame = new Mat(inputFrame.rgba().size(), inputFrame.rgba().type());
-        Imgproc.undistort(inputFrame.rgba(), renderedFrame,
+        Calib3d.undistort(inputFrame.rgba(), renderedFrame,
                 mCalibrator.getCameraMatrix(), mCalibrator.getDistortionCoefficients());
 
         return renderedFrame;
@@ -71,7 +72,7 @@ class ComparisonFrameRender extends FrameRender {
     @Override
     public Mat render(CvCameraViewFrame inputFrame) {
         Mat undistortedFrame = new Mat(inputFrame.rgba().size(), inputFrame.rgba().type());
-        Imgproc.undistort(inputFrame.rgba(), undistortedFrame,
+        Calib3d.undistort(inputFrame.rgba(), undistortedFrame,
                 mCalibrator.getCameraMatrix(), mCalibrator.getDistortionCoefficients());
 
         Mat comparisonFrame = inputFrame.rgba();
